// src/app/models/DelegationData.ts
export interface DelegationData {
  delegationNr: string;
  name: string;
  date: string;
  destination: string;
  delegationPlace: string;
  costs: string;
  advancePay: string;
  transport: string;
  /* Production version:
  registrationNr: number;
  delegationNr: number;
  mileage: number;
  costs: string; */
}